moved serialization code back in ViewerData.h
